move supplier allocation table to seperate component
