+++ admin-portal/src/pages/aidPackage/assignSuppliers/assignSuppliers.tsx
@@ -1,7 +1,8 @@
-import { NeedsInfo, SupplierQuote } from "data/medical-needs.mock.data";
-import { useState, useEffect, useMemo } from "react";
+import { NeedsInfo } from "data/medical-needs.mock.data";
+import { useMemo } from "react";
 import { useTable, useExpanded } from "react-table";
 import { NeedAssignments } from "../aidPackage";
+import { SupplierNeedAllocationTable } from "./supplierNeedAllocationTable";
 
 export function AssignSuppliers({
   needAssignments,
@@ -32,10 +33,10 @@ export function AssignSuppliers({
         );
         return {
           needID: need.needID,
-          need: need.name,
+          needName: need.name,
           period: periodDate.toLocaleDateString(),
-          qtyNeeded: need.remainingQuantity.toString(),
-          remainingNeed:
+          requiredQuantity: need.remainingQuantity,
+          remainingQuantity:
             need.remainingQuantity - getAssignedCount(need, needAssignments),
           supplierQuotes: need.supplierQuotes || [],
           expanded: true,
@@ -69,11 +70,11 @@ export function AssignSuppliers({
       },
       {
         Header: "Quantity Needed",
-        accessor: "qtyNeeded",
+        accessor: "requiredQuantity",
       },
       {
         Header: "Remaining Need",
-        accessor: "remainingNeed",
+        accessor: "remainingQuantity",
       },
     ],
     []
@@ -89,7 +90,7 @@ export function AssignSuppliers({
   } = useTable({ columns, data, autoResetHiddenColumns: false }, useExpanded);
 
   return (
-    <table {...getTableProps()} style={{ border: "solid 1px blue" }}>
+    <table {...getTableProps()}>
       <thead>
         {headerGroups.map((headerGroup) => (
           <tr {...headerGroup.getHeaderGroupProps()}>
@@ -103,12 +104,13 @@ export function AssignSuppliers({
       </thead>
 
       <tbody {...getTableBodyProps()}>
+        {/* TODO: add validation for remaining need */}
         {rows.map((row, index) => {
           prepareRow(row);
 
           const needsID = row.original["needID"];
           const currentAssignments = needAssignments[needsID];
-          const remainingNeed = row.values["remainingNeed"];
+          const requiredQuantity = row.values["requiredQuantity"];
 
           return (
             <>
@@ -148,7 +150,7 @@ export function AssignSuppliers({
                           [needsID]: updatedAssignments,
                         });
                       }}
-                      remainingNeed={remainingNeed}
+                      requiredQuantity={requiredQuantity}
                     />
                   </td>
                 </tr>
@@ -159,108 +161,4 @@ export function AssignSuppliers({
       </tbody>
     </table>
   );
-}
-
-function SupplierNeedAllocationTable({
-  supplierQuotes,
-  setAssignmentForSupplier,
-  remainingNeed,
-  assignmentsForSupplier,
-}: {
-  supplierQuotes: SupplierQuote[];
-  setAssignmentForSupplier: any;
-  remainingNeed: number;
-  assignmentsForSupplier: Map<number, number>;
-}) {
-  supplierQuotes = supplierQuotes || [];
-
-  const data = useMemo<Array<{ [key: string]: any }>>(
-    () =>
-      supplierQuotes.map((quote) => ({
-        supplier: quote.supplier.name,
-        quantity: assignmentsForSupplier.get(quote.supplierID),
-        max: Math.min(remainingNeed, quote.availableQuantity),
-        supplierID: quote.supplierID,
-      })),
-    [supplierQuotes, assignmentsForSupplier, remainingNeed]
-  );
-
-  const columns = useMemo(
-    () => [
-      {
-        Header: "Supplier",
-        accessor: "supplier",
-      },
-      {
-        Header: "Max",
-        accessor: "max",
-      },
-      {
-        Header: "Order Quantity",
-        accessor: "quantity",
-        Cell: EditableCell,
-      },
-    ],
-    []
-  );
-
-  const { getTableProps, getTableBodyProps, headerGroups, rows, prepareRow } =
-    useTable<any>({
-      columns,
-      data,
-      autoResetHiddenColumns: false,
-      // @ts-ignore
-      updateAssignment: (index: number, id: string, value: number) => {
-        const rowValues = rows[index].original;
-        setAssignmentForSupplier(rowValues.supplierID, value);
-      },
-    });
-
-  return (
-    <table {...getTableProps()}>
-      <thead>
-        {headerGroups.map((headerGroup) => (
-          <tr {...headerGroup.getHeaderGroupProps()}>
-            {headerGroup.headers.map((column) => (
-              <th {...column.getHeaderProps()}>{column.render("Header")}</th>
-            ))}
-          </tr>
-        ))}
-      </thead>
-      <tbody {...getTableBodyProps()}>
-        {rows.map((row, i) => {
-          prepareRow(row);
-          return (
-            <tr {...row.getRowProps()}>
-              {row.cells.map((cell) => {
-                return <td {...cell.getCellProps()}>{cell.render("Cell")}</td>;
-              })}
-            </tr>
-          );
-        })}
-      </tbody>
-    </table>
-  );
-}
-
-function EditableCell({
-  value: initialValue,
-  row: { index },
-  column: { id },
-  updateAssignment, // This is a custom function that we supplied to our table instance
-}: any) {
-  // We need to keep and update the state of the cell normally
-  const [value, setValue] = useState(initialValue);
-
-  const onChange = (e: any) => {
-    setValue(e.target.value);
-    updateAssignment(index, id, e.target.value);
-  };
-
-  // If the initialValue is changed external, sync it up with our state
-  useEffect(() => {
-    setValue(initialValue);
-  }, [initialValue]);
-
-  return <input type="number" value={value} onChange={onChange} />;
 }
